+++ lib/sdui.dart
@@ -16,6 +16,7 @@ export 'src/input.dart';
 export 'src/list_view.dart';
 export 'src/page_view.dart';
 export 'src/parser.dart';
+export 'src/pin_with_keyboard.dart';
 export 'src/radio.dart';
 export 'src/route.dart';
 export 'src/row.dart';