+++ lib/src/action.dart
@@ -16,14 +16,15 @@ typedef ActionCallback = Future<String?> Function(BuildContext context);
 ///
 /// ### JSON Attributes
 /// - **type**: Type of action. The supported values are:
-///   - `screen`: To redirect to another screen
-///   - `page`: To redirect to another page, in the context of [PageView]
-///   - `command`: To execute a command on the server
-/// - **url**: Action URL
-///   - ``route:/..``: redirect users to previous route
-///   - ``page:/<PAGE_NUMBER>``: redirect users to a given page. `<PAGE_NUMBER>` is the page index (starting with `0`).
-///   - URL starting with ``route:/<ROUTE_NAME>`` redirect user the a named route. (Ex: ``route:/checkout``)
-///   - URL starting with ``http://`` or ``https`` redirect user to a server driven page
+///   - `Route`: To redirect to another route
+///   - `Page`: To redirect to another page, in the context of [PageView]
+///   - `Command`: To execute a command on the server
+/// - **url**: Action URL. This URL represent either the route or a command to execute
+///   - `route:/..`: redirect users to previous route
+///   - URL starting with `route:/<ROUTE_NAME>` redirect user the a named route. (Ex: ``route:/checkout``)
+///   - URL starting with `http://` or `https` redirect user to a server driven page
+///   - `page:/<PAGE_NUMBER>`: redirect users to a given page. `<PAGE_NUMBER>` is the page index (starting with `0`).
+/// - **replacement**: For `type=route`, this indicate if we replace the current view or navigate.
 class SDUIAction {
   static final Logger _logger = Logger();
 
@@ -31,6 +32,7 @@ class SDUIAction {
 
   String? type;
   String url = '';
+  bool replacement = false;
   SDUIDialog? prompt;
 
   /// controller associated with the action
@@ -39,6 +41,7 @@ class SDUIAction {
   SDUIAction fromJson(Map<String, dynamic>? attributes) {
     url = attributes?["url"] ?? '';
     type = attributes?["type"];
+    replacement = attributes?["replacement"] ?? false;
 
     var prompt = attributes?["prompt"];
     if (prompt is Map<String, dynamic>) {
@@ -51,15 +54,17 @@ class SDUIAction {
           BuildContext context, Map<String, dynamic>? data) async =>
       _prompt(context).then((value) => _execute(value, context, data));
 
-  Future<String> _execute(
+  Future<String?> _execute(
       String? result, BuildContext context, Map<String, dynamic>? data) async {
     if (result == null) {
       return _emptyFuture;
     }
 
+    _logger.i('Executing: type=$type url=$url');
+
     switch (type?.toLowerCase()) {
-      case 'screen':
-        return _gotoScreen(context, data);
+      case 'route':
+        return _gotoRoute(context, data);
 
       case 'page':
         return _gotoPage(context, data);
@@ -94,23 +99,36 @@ class SDUIAction {
     return _emptyFuture;
   }
 
-  Future<String> _gotoScreen(BuildContext context, Map<String, dynamic>? data) {
+  Future<String> _gotoRoute(BuildContext context, Map<String, dynamic>? data) {
     if (_isRoute()) {
       _logger.i('Navigating to route $url');
       var route = url.substring(6);
       if (route == '/..') {
         Navigator.pop(context);
       } else {
-        Navigator.pushNamed(context, route);
+        if (replacement) {
+          Navigator.pushReplacementNamed(context, route);
+        } else {
+          Navigator.pushNamed(context, route);
+        }
       }
     } else if (_isNetwork()) {
       _logger.i('Navigating to screen $url');
-      Navigator.push(
-        context,
-        MaterialPageRoute(
-            builder: (context) =>
-                DynamicRoute(provider: HttpRouteContentProvider(url))),
-      );
+      if (replacement) {
+        Navigator.pushReplacement(
+          context,
+          MaterialPageRoute(
+              builder: (context) =>
+                  DynamicRoute(provider: HttpRouteContentProvider(url))),
+        );
+      } else {
+        Navigator.push(
+          context,
+          MaterialPageRoute(
+              builder: (context) =>
+                  DynamicRoute(provider: HttpRouteContentProvider(url))),
+        );
+      }
     }
     return _emptyFuture;
   }